image upload middleware add condition for no error thrw when no image uploaded

+++ imageuploadmiddleware/multipleUploadbySeperate.js
@@ -33,55 +33,59 @@ const uploadBySeperate = (req, res, next) => {
             return res.status(400).json({ error: err.message })
         }
 
-        const files1 = req.files[field1]
-        const files2 = req.files[field2]
-        const files3 = req.files[field3]
-        const files4 = req.files[field4]
-        const files5 = req.files[field5]
+        if (req.files) {
+            const files1 = req.files[field1]
+            const files2 = req.files[field2]
+            const files3 = req.files[field3]
+            const files4 = req.files[field4]
+            const files5 = req.files[field5]
 
-        const errors = [];
+            const errors = [];
 
-        console.log("ddwq", files5);
-        [files1, files2, files3, files4, files5].forEach((files, index) => {
+            console.log("ddwq", files5);
+            [files1, files2, files3, files4, files5].forEach((files, index) => {
 
-            const allowedTypes = ['image/jpeg', 'image/png', 'video/mp4'];
-            const maxSize = 10 * 1024 * 1024  // 10MB
+                const allowedTypes = ['image/jpeg', 'image/png', 'video/mp4'];
+                const maxSize = 10 * 1024 * 1024  // 10MB
 
-            if (!files || files.length === 0) {
-                errors.push(`No file provided for ${getFieldName(index)}`);
-            } else {
+                if (!files || files.length === 0) {
+                    errors.push(`No file provided for ${getFieldName(index)}`);
+                } else {
 
-                const file = files[0];
+                    const file = files[0];
 
-                if (!allowedTypes.includes(file.mimetype)) {
-                    errors.push(`Invalid file type for ${getFieldName(index)} : ${file.originalname}`);
+                    if (!allowedTypes.includes(file.mimetype)) {
+                        errors.push(`Invalid file type for ${getFieldName(index)} : ${file.originalname}`);
+                    }
+                    if (file.size > maxSize) {
+                        errors.push(`Invalid file type for ${getFieldName(index)} : ${file.originalname}`);
+                    }
                 }
-                if (file.size > maxSize) {
-                    errors.push(`Invalid file type for ${getFieldName(index)} : ${file.originalname}`);
-                }
+            });
+
+            if (errors.length > 0) {
+
+                [files1, files2, files3, files4, files5].forEach((files) => {
+                    if (files && files.length > 0) {
+                        fs.unlinkSync(files[0].path)
+                    }
+                })
+
+                return res.status(400).json({ errors });
             }
-        });
 
-        if (errors.length > 0) {
+            req.files = {
+                [field1]: files1 ? files1[0] : null,
+                [field2]: files2 ? files2[0] : null,
+                [field3]: files3 ? files3[0] : null,
+                [field4]: files4 ? files4[0] : null,
+                [field5]: files5 ? files5[0] : null
+            }
 
-            [files1, files2, files3, files4, files5].forEach((files) => {
-                if (files && files.length > 0) {
-                    fs.unlinkSync(files[0].path)
-                }
-            })
-
-            return res.status(400).json({ errors });
+            next();
+        } else {
+            next();
         }
-
-        req.files = {
-            [field1]: files1 ? files1[0] : null,
-            [field2]: files2 ? files2[0] : null,
-            [field3]: files3 ? files3[0] : null,
-            [field4]: files4 ? files4[0] : null,
-            [field5]: files5 ? files5[0] : null
-        }
-
-        next();
     })
 }
